Implementing hierarchial softmax

+++ trainer/trainer.py
@@ -221,7 +221,7 @@ class Trainer(GenericTrainer):
 
             y_onehot.zero_()
             output = self.model(Variable(data),predictClass=True)
-            target = (target/self.args.step_size).int().float()
+            target = (target/self.args.step_size).int().long()
             target.unsqueeze_(1)
             y_onehot.scatter_(1, target, 1)
             lossHigher = F.kl_div(output, Variable(y_onehot))
